\name{db.disconnect}
\alias{db.disconnect}
\title{End connection to postgres backend}
\usage{
db.disconnect()
}
\description{
  db.disconnect disconnects from the currently open database.
}
\references{\url{http://www.postgresql.org/}}
\author{Timothy H. Keitt}
\seealso{\code{\link{db.connect()}}}
\keyword{database}

