\name{db.error.message}
\alias{db.error.message}
\title{Get current error message from libpq}
\usage{
db.error.message()
}
\description{
  Returns the last error message generated in libpq.
}
\value{
  \item{message}{A character string containing the error message}
}
\author{Timothy H. Keitt}
\keyword{database}
