\name{list.to.csv}
\alias{list.to.csv}
\title{Convert a list of values to a string containing comma separated values}
\usage{
list.to.csv(...)
}
\arguments{
 \item{\dots}{All arguments are passed to \code{\link{paste}}}
}
\description{
  Convert a list of values to a string containing comma separated values
}
\value{
  A character string with comma separated values
}
\author{Timothy H. Keitt}
\keyword{database}
