\name{pgSQLConnection-methods}
\docType{methods}
\alias{pgSQLConnection-methods}
\alias{dbWriteTable,pgSQLConnection-method}
\alias{dbDataType,pgSQLConnection-method}
\alias{dbSendQuery,pgSQLConnection,character-method}
\title{Methods for the class `pgSQLConnect' in Package `pgSQL'}
\description{
  Methods for the class `pgSQLConnection' in Package `pgSQL'.  
}
\section{Methods}{
  \describe{
    \item{dbWriteTable}{\code{signature(conn = "pgSQLConnection", ...)}}

}}
\keyword{methods}

