\name{psql}
\alias{psql}
\title{Run the PostgreSQL monitor}
\description{
  \code{psql} simply runs the PostgreSQL monitor application "psql".
}
\usage{
psql(dbname=NULL, host=NULL, port=NULL)
}
\arguments{
  \item{dbname}{a database name}
  \item{host}{a database host name}
  \item{port}{a port number}
}
\author{Timothy H. Keitt}
\keyword{database}
