\name{to.R.name}
\alias{to.R.name}
\title{Convert a postgres name to an R name}
\usage{
to.R.name(name)
}
\arguments{
 \item{name}{A character string containing the R object name}
}
\description{
  \code{to.R.name} converts all underscores to dots "."
}
\value{
  An R freindly name
}
\author{Timothy H. Keitt}
\seealso{\code{\link{to.db.name}}}
\keyword{database}
